requirements.txt
boto3==1.14.*
cloudevents==1.2.*
Flask==1.1.*
flask-cors==3.0.*
kafka-python==2.0.2
numpy
